Refactor test suites
