+++ app/src/androidTest/java/com/andela/movit/background/ServiceHelperTest.java
@@ -1,0 +1,7 @@
+package com.andela.movit.background;
+
+import static org.junit.Assert.*;
+
+public class ServiceHelperTest {
+
+}
