+++ app/src/main/java/com/andela/movit/activityrecognition/ActivityRecognitionService.java
@@ -53,10 +53,7 @@ public class ActivityRecognitionService extends IntentService {
         return new Comparator<DetectedActivity>() {
             @Override
             public int compare(DetectedActivity lhs, DetectedActivity rhs) {
-                if (rhs.getConfidence() > lhs.getConfidence()) {
-                    return 1;
-                }
-                return 0;
+                return rhs.getConfidence() - lhs.getConfidence();
             }
         };
     }
