+++ app/src/main/java/com/andela/movit/utilities/Utility.java
@@ -21,66 +21,10 @@ import static com.andela.movit.config.Constants.*;
 
 public class Utility {
 
-    public static Movit getApp() {
-        return Movit.getApp();
-    }
-
-    public static void startActivity(Context context, Class activityClass) {
-        Intent intent = new Intent(context, activityClass);
-        context.startActivity(intent);
-    }
-
-    public static void makeToast(Context context, String message) {
-        Toast.makeText(context, message, Toast.LENGTH_SHORT).show();
-    }
-
-    public static void stopService(Context context, Class serviceClass) {
-        context.stopService(new Intent(context, serviceClass));
-    }
-
-    public static Intent putMovementInIntent(Movement movement, Intent intent) {
-        intent.putExtra(PLACE_NAME.getValue(), movement.getPlaceName());
-        intent.putExtra(ACTIVITY_NAME.getValue(), movement.getActivityName());
-        intent.putExtra(LATITUDE.getValue(), Double.toString(movement.getLatitude()));
-        intent.putExtra(LONGITUDE.getValue(), Double.toString(movement.getLongitude()));
-        intent.putExtra(TIMESTAMP.getValue(), Long.toString(movement.getTimeStamp()));
-        return intent;
-    }
-
-    public static Movement getMovementFromBundle(Bundle bundle) {
-        Movement mv = new Movement();
-        mv.setActivityName(getStringFromBundle(bundle, ACTIVITY_NAME.getValue()));
-        mv.setPlaceName(getStringFromBundle(bundle, PLACE_NAME.getValue()));
-        mv.setLatitude(Double.parseDouble(getStringFromBundle(bundle, LATITUDE.getValue())));
-        mv.setLongitude(Double.parseDouble(getStringFromBundle(bundle, LONGITUDE.getValue())));
-        mv.setTimeStamp(Long.parseLong(getStringFromBundle(bundle, TIMESTAMP.getValue())));
-        return mv;
-    }
-
-    private static String getStringFromBundle(Bundle bundle, String name) {
-        return bundle.getString(name);
-    }
-
     public static String getCoordsString(Movement movement) {
         return Double.toString(movement.getLatitude())
                 + ", "
                 + Double.toString(movement.getLongitude());
-    }
-
-    public static StringBroadcastReceiver registerStringReceiver(Context context, String actionName) {
-        StringBroadcastReceiver receiver = new StringBroadcastReceiver(actionName);
-        LocalBroadcastManager
-                .getInstance(context)
-                .registerReceiver(receiver, getFilter(actionName));
-        return receiver;
-    }
-
-    private static IntentFilter getFilter(String actionName) {
-        return new IntentFilter(actionName);
-    }
-
-    public static void unregisterReceiver(Context context, BroadcastReceiver receiver) {
-        LocalBroadcastManager.getInstance(context).unregisterReceiver(receiver);
     }
 
     public static int[] getDateValues(Date date) {
@@ -104,10 +48,6 @@ public class Utility {
         return format.format(date);
     }
 
-    public static LayoutInflater getInflater(Context context) {
-        return (LayoutInflater) context.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
-    }
-
     public static String getDurationString(long duration) {
         long longVal = duration / 1000;
         int hours = (int) longVal / 3600;
@@ -121,9 +61,5 @@ public class Utility {
     private static String formatNumber(int number) {
         String numString = Integer.toString(number);
         return number < 9 ? "0" + numString : numString;
-    }
-
-    public static void broadcastIntent(Context context, Intent intent) {
-        LocalBroadcastManager.getInstance(context).sendBroadcast(intent);
     }
 }
